Preparação para o Heroku

+++ src/main/java/com/joao/hotwheelsapi/domain/repository/CarrinhoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.joao.hotwheelsapi.domain.model.Carrinho;
 
-public interface CarrinhoRepository extends JpaRepository<Carrinho, Integer>{
+public interface CarrinhoRepository extends JpaRepository<Carrinho, Long>{
 
 }
